split out curve from line
