+++ examples/amr_volume_rendering_with_curves.py
@@ -1,7 +1,7 @@
-import yt
 import numpy as np
-import yt_idv
+import yt
 
+import yt_idv
 from yt_idv.scene_components.curves import CurveRendering  # NOQA
 from yt_idv.scene_data.line import CurveData  # NOQA
